server: Recover from panicking handlers
A single panic() inside a plugin causes the entire server to crash.
Since panics are rarely used on purpose in modern go and usually runtime
errors, this is maybe too violent, especially as requests are
independent from each other.

This uses recover() to catch panics. When a handler panics, the request
is dropped but the server continues to handle requests.

Signed-off-by: Anatole Denis <anatole@unverle.fr>